Cache for command results
Adds a HashMap of type String,Movie that saves a command as a key and
saves the result as a value. The server checks when a user enters a
command if its been entered before and if so returns that querys result
from the hashmap instead of querying the database again.